% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{calc_stats}
\alias{calc_stats}
\title{Simple spatial statistics for users}
\usage{
calc_stats(mobility_data, spatial_zones)
}
\arguments{
\item{mobility_data}{Mobility data}

\item{spatial_zones}{Spatial zones sf object}
}
\value{
Data frame with spatial statistics
}
\description{
Calculate basic spatial statistics that can be easily mapped and understood
by users.
}
\examples{
\dontrun{
# Calculate spatial statistics
spatial_stats <- calc_stats(mobility_data, zones)

# Create map of spatial statistics
map <- create_spatial_map(spatial_stats, zones, map_type = "choropleth")
}
}
