% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{calculate_stats}
\alias{calculate_stats}
\title{Create spatial statistics summary}
\usage{
calculate_stats(mobility_data, spatial_zones)
}
\arguments{
\item{mobility_data}{Mobility data}

\item{spatial_zones}{sf object with spatial zones}
}
\value{
Data frame with spatial statistics
}
\description{
Calculate basic spatial statistics for mobility data and zones.
}
\keyword{internal}
