% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{create_flow_map_sf}
\alias{create_flow_map_sf}
\title{Create flow map using sf}
\usage{
create_flow_map_sf(analysis_result, spatial_zones, title)
}
\arguments{
\item{analysis_result}{Analysis result data}

\item{spatial_zones}{Spatial zones sf object}

\item{title}{Map title}
}
\value{
ggplot2 flow map
}
\description{
Create flow map using sf
}
\keyword{internal}
