% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{create_point_map_sf}
\alias{create_point_map_sf}
\title{Create point map using sf}
\usage{
create_point_map_sf(analysis_result, spatial_zones, color_palette, title)
}
\arguments{
\item{analysis_result}{Analysis result data}

\item{spatial_zones}{Spatial zones sf object}

\item{color_palette}{Color palette}

\item{title}{Map title}
}
\value{
ggplot2 point map
}
\description{
Create point map using sf
}
\keyword{internal}
