% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{create_simple_prediction_data}
\alias{create_simple_prediction_data}
\title{Create simple prediction data}
\usage{
create_simple_prediction_data(prediction_dates)
}
\arguments{
\item{prediction_dates}{Dates to predict}
}
\value{
Simple prediction data frame
}
\description{
Create simple prediction data
}
\keyword{internal}
