% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{detect_map_type}
\alias{detect_map_type}
\title{Detect appropriate map type based on analysis result}
\usage{
detect_map_type(analysis_result)
}
\arguments{
\item{analysis_result}{Analysis result data}
}
\value{
Character string indicating map type
}
\description{
Detect appropriate map type based on analysis result
}
\keyword{internal}
