% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{detect_value_column}
\alias{detect_value_column}
\title{Detect the main value column for mapping}
\usage{
detect_value_column(analysis_result)
}
\arguments{
\item{analysis_result}{Analysis result data}
}
\value{
Character string indicating value column
}
\description{
Detect the main value column for mapping
}
\keyword{internal}
