% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{get_city_characteristics}
\alias{get_city_characteristics}
\title{Get city characteristics for mobility analysis}
\usage{
get_city_characteristics(city_name)
}
\arguments{
\item{city_name}{Character. Name of the city}
}
\value{
List with city characteristics
}
\description{
Get city characteristics for mobility analysis
}
\keyword{internal}
