% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{mobspain}
\alias{mobspain-package}
\alias{mobspain}
\title{mobspain: Spanish Transport Zone Analysis}
\description{
A minimal R package for accessing Spanish administrative zones
for transport analysis. Provides easy access to spatial zone
geometries with filtering capabilities.
}
\details{
The package contains one core function:
\itemize{
\item \code{\link{get_zones}} - Download Spanish administrative zones
}
}
\author{
Prince Oppong Boakye
}
