% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{print.mobspain_anomalies}
\alias{print.mobspain_anomalies}
\title{Print method for mobspain_anomalies objects}
\usage{
\method{print}{mobspain_anomalies}(x, ...)
}
\arguments{
\item{x}{mobspain_anomalies object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for mobspain_anomalies objects
}
