% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{print.mobspain_containment}
\alias{print.mobspain_containment}
\title{Print method for mobspain_containment objects}
\usage{
\method{print}{mobspain_containment}(x, ...)
}
\arguments{
\item{x}{mobspain_containment object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for mobspain_containment objects
}
