% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{print.simple_predictions}
\alias{print.simple_predictions}
\title{Print method for simple predictions}
\usage{
\method{print}{simple_predictions}(x, ...)
}
\arguments{
\item{x}{Predictions object}

\item{...}{Additional arguments}
}
\description{
Print method for simple predictions
}
