% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_zones}
\alias{sample_zones}
\title{Sample Spatial Zones Dataset}
\format{
A data frame (sf object) with 3 rows and 5 variables:
\describe{
\item{id}{Zone identifier (character)}
\item{name}{Zone name (character)}
\item{population}{Population count (numeric)}
\item{geometry}{Spatial geometry (sf geometry column)}
\item{area_km2}{Area in square kilometers (numeric)}
}
}
\source{
Generated for demonstration purposes
}
\usage{
sample_zones
}
\description{
A sample dataset containing spatial zones with geographic boundaries for
demonstration and testing purposes in the mobspain package.
}
\details{
This dataset provides example spatial zones with point geometries that can be
used for testing spatial analysis functions in the mobspain package. The zones
include basic attributes like population and calculated area.
}
\examples{
data(sample_zones)

# View the structure
str(sample_zones)

# Plot the zones
if(require(sf)) {
  plot(sample_zones$geometry)
}

}
\keyword{datasets}
