% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{summary.mobspain_anomalies}
\alias{summary.mobspain_anomalies}
\title{Summary method for mobspain_anomalies objects}
\usage{
\method{summary}{mobspain_anomalies}(object, ...)
}
\arguments{
\item{object}{mobspain_anomalies object}

\item{...}{Additional arguments (ignored)}
}
\description{
Summary method for mobspain_anomalies objects
}
