% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{summary.mobspain_containment}
\alias{summary.mobspain_containment}
\title{Summary method for mobspain_containment objects}
\usage{
\method{summary}{mobspain_containment}(object, ...)
}
\arguments{
\item{object}{mobspain_containment object}

\item{...}{Additional arguments (ignored)}
}
\description{
Summary method for mobspain_containment objects
}
