% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{train_simple_linear_model}
\alias{train_simple_linear_model}
\title{Train a simple linear regression model}
\usage{
train_simple_linear_model(X_train, y_train)
}
\arguments{
\item{X_train}{Training features}

\item{y_train}{Training target}
}
\value{
Trained linear model
}
\description{
Train a simple linear regression model
}
\keyword{internal}
