% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{train_simple_random_forest}
\alias{train_simple_random_forest}
\title{Train a simple random forest model}
\usage{
train_simple_random_forest(X_train, y_train)
}
\arguments{
\item{X_train}{Training features}

\item{y_train}{Training target}
}
\value{
Trained random forest model
}
\description{
Train a simple random forest model
}
\keyword{internal}
